--- torchtext/nn/modules/__init__.py
@@ -1,6 +1,0 @@
-from .multiheadattention import InProjContainer, \
-    MultiheadAttentionContainer, ScaledDotProduct
-
-__all__ = ['InProjContainer',
-           'MultiheadAttentionContainer',
-           'ScaledDotProduct']
